# medienverwaltungweb.client/swagger.json
{
    "openapi": "3.0.4",
    "info": {
      "title": "medienVerwaltungWeb.Server",
      "version": "1.0"
    },
    "paths": {
...
      "/api/Movie": {
        "get": {
          "tags": [
            "Movie"
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/MovieDTO"
                    }
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/MovieDTO"
                    }
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/MovieDTO"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Movie"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieDTO"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieDTO"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/MovieDTO"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/MovieDTO"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/MovieDTO"
                  }
                }
              }
            }
          }
        }
      },
      "/api/Movie/{id}": {
        "get": {
          "tags": [
            "Movie"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/MovieDTO"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/MovieDTO"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/MovieDTO"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "Movie"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieDTO"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieDTO"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "delete": {
          "tags": [
            "Movie"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
...
      "/api/MusicAlbum": {
        "get": {
          "tags": [
            "MusicAlbum"
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/MusicAlbumDTO"
                    }
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/MusicAlbumDTO"
                    }
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/MusicAlbumDTO"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "MusicAlbum"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MusicAlbumDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MusicAlbumDTO"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/MusicAlbumDTO"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/MusicAlbumDTO"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/MusicAlbumDTO"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/MusicAlbumDTO"
                  }
                }
              }
            }
          }
        }
      },
      "/api/MusicAlbum/{id}": {
        "get": {
          "tags": [
            "MusicAlbum"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/MusicAlbumDTO"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/MusicAlbumDTO"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/MusicAlbumDTO"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "MusicAlbum"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MusicAlbumDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MusicAlbumDTO"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/MusicAlbumDTO"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "delete": {
          "tags": [
            "MusicAlbum"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "MovieDTO": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "title": {
              "type": "string",
              "nullable": true
            },
            "length": {
              "type": "integer",
              "format": "int32"
            },
            "releaseYear": {
              "type": "integer",
              "format": "int32"
            },
            "description": {
              "type": "string",
              "nullable": true
            },
            "genre": {
              "type": "string",
              "nullable": true
            },
            "actorIDs": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              },
              "nullable": true
            },
            "location": {
              "type": "string",
              "nullable": true
            },
            "actorCountDisplay": {
              "type": "string",
              "nullable": true,
              "readOnly": true
            }
          },
          "additionalProperties": false
        },
        "MusicAlbumDTO": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "title": {
              "type": "string",
              "nullable": true
            },
            "songIdList": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              },
              "nullable": true
            },
            "interpretFullName": {
              "type": "string",
              "nullable": true
            },
            "location": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
...
      }
    }
  }